+++ lib/modules/datasource/azure-pipelines-tasks/schema.ts
